refactor(chat): 인프라 설정을 AppModule로 분리
